[refactor] #60 change vlock creation from async to sync
