+++ src/main/java/org/umc/travlocksserver/domain/vlock/controller/VlockController.java
@@ -11,9 +11,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import org.umc.travlocksserver.domain.vlock.constant.VlockSuccessCode;
-import org.umc.travlocksserver.domain.vlock.dto.vlock.VlockRequestDTO;
-import org.umc.travlocksserver.domain.vlock.dto.vlock.VlockResponseDTO;
+import org.umc.travlocksserver.domain.vlock.code.VlockSuccessCode;
+import org.umc.travlocksserver.domain.vlock.dto.request.VlockRequestDTO;
+import org.umc.travlocksserver.domain.vlock.dto.response.VlockResponseDTO;
 import org.umc.travlocksserver.domain.vlock.service.VlockCommandService;
 import org.umc.travlocksserver.domain.vlock.service.VlockQueryService;
 import org.umc.travlocksserver.global.response.SuccessResponse;
@@ -31,14 +31,14 @@ public class VlockController implements VlockControllerDocs {
 
 	/** 블록 생성 */
 	@PostMapping
-	public ResponseEntity<SuccessResponse<Void>> createVlock(
+	public ResponseEntity<SuccessResponse<VlockResponseDTO>> createVlock(
 		@AuthenticationPrincipal Long memberId,
 		@Valid @RequestBody VlockRequestDTO request) {
-		vlockCommandService.createVlock(memberId, request);
+		VlockResponseDTO response = vlockCommandService.createVlock(memberId, request);
 
 		return ResponseEntity
-			.status(HttpStatus.ACCEPTED)
-			.body(SuccessResponse.ok(VlockSuccessCode.VLOCK_CREATE_ACCEPTED_SUCCESS));
+			.status(HttpStatus.CREATED)
+			.body(SuccessResponse.ok(VlockSuccessCode.VLOCK_CREATE_SUCCESS, response));
 	}
 
 	/** 인기 블록 조회 */
