+++ src/main/java/org/umc/travlocksserver/domain/vlock/repository/VlockRepositoryCustom.java
@@ -2,7 +2,7 @@ package org.umc.travlocksserver.domain.vlock.repository;
 
 import java.util.List;
 
-import org.umc.travlocksserver.domain.vlock.dto.vlock.VlockResponseDTO;
+import org.umc.travlocksserver.domain.vlock.dto.response.VlockResponseDTO;
 
 public interface VlockRepositoryCustom {
