+++ src/main/java/org/umc/travlocksserver/domain/vlock/service/VlockCommandService.java
@@ -10,11 +10,14 @@ import org.umc.travlocksserver.domain.member.entity.Member;
 import org.umc.travlocksserver.domain.member.exception.MemberException;
 import org.umc.travlocksserver.domain.member.exception.code.MemberErrorCode;
 import org.umc.travlocksserver.domain.member.repository.MemberRepository;
-import org.umc.travlocksserver.domain.vlock.dto.vlock.VlockRequestDTO;
+import org.umc.travlocksserver.domain.vlock.dto.request.VlockRequestDTO;
+import org.umc.travlocksserver.domain.vlock.dto.response.VlockResponseDTO;
+import org.umc.travlocksserver.domain.vlock.entity.Vlock;
 import org.umc.travlocksserver.domain.vlock.entity.VlockCategory;
 import org.umc.travlocksserver.domain.vlock.exception.VlockException;
-import org.umc.travlocksserver.domain.vlock.constant.VlockErrorCode;
+import org.umc.travlocksserver.domain.vlock.code.VlockErrorCode;
 import org.umc.travlocksserver.domain.vlock.repository.VlockCategoryRepository;
+import org.umc.travlocksserver.domain.vlock.repository.VlockRepository;
 
 import lombok.RequiredArgsConstructor;
 
@@ -26,9 +29,9 @@ public class VlockCommandService {
 	private final VlockCategoryRepository vlockCategoryRepository;
 	private final CityRepository cityRepository;
 	private final MemberRepository memberRepository;
-	private final VlockAsyncHandler vlockAsyncHandler;
+	private final VlockRepository vlockRepository;
 
-	public void createVlock(Long memberId, VlockRequestDTO request) {
+	public VlockResponseDTO createVlock(Long memberId, VlockRequestDTO request) {
 		Member member = memberRepository.findById(memberId)
 			.orElseThrow(() -> new MemberException(MemberErrorCode.MEMBER_NOT_FOUND));
 
@@ -38,7 +41,20 @@ public class VlockCommandService {
 		City city = cityRepository.findWithRegionById(request.cityId())
 			.orElseThrow(() -> new CityException(CityErrorCode.CITY_NOT_FOUND));
 
-		vlockAsyncHandler.saveVlockAsync(member, category, city, request);
+		Vlock newVlock = Vlock.create(
+			category,
+			city,
+			member,
+			request.name(),
+			request.latitude(),
+			request.longitude(),
+			request.address(),
+			request.memo()
+		);
+
+		Vlock savedVlock = vlockRepository.save(newVlock);
+
+		return VlockResponseDTO.from(savedVlock);
 	}
 }
